requirements.txt
Flask==2.3.2
celery==5.3.1
redis==4.5.5
Pillow==9.5.0
requests==2.31.0